refactor: Line 값객체 도입
